requirements.txt
# Core pipeline dependencies
pandas>=1.3.0
pyodbc>=4.0.0
sqlalchemy>=1.4.0
google-cloud-bigquery>=2.0.0
google-auth>=1.30.0
python-dotenv>=0.19.0

# Testing dependencies
pytest>=6.0.0
pytest-mock>=3.0.0
ipython>=7.0.0  # Pour %run magic command